Platform: dStrtod
String to double:
Skips whitespace until it can scan something (important,) and tries to scan a number next. The returned double is like atod, but endptr is a pointer to the start of the text if the double failed to parse, conversely it's at the end if a double could be parsed.